Mixed Entries Comparison works now. Test: packet 1: [5] packet 2: [[8,7,6]] packet 1 < packet 2 successful.
The reason is that Ord does not replace the <, >, = operators.
when testing,
I assumed that the Ord cmp method replaces those operators
but it does not

PartialOrd is the struct that can replace <, >, =, ...

but we were deriving the default implementation of PartialOrd
for our custom nested PacketDatum struct.

Instead we had to:
implement our own PartialOrd & use our custom cmp method from impl Ord